- Move ImageData::save from protected to public - Created the Pixel class to replace PixelRef - Refactored ImageData to use Pixel
Note: I've made these changes because it is a more rational design that the use of PixelRef. Now Pixels can have an interface that allows for easy per-pixel operations/filters etc. Pixel also works with floats instead of unsigned chars and accessing an manipulating them is far simpler this way.